Move utilities providing dialect information into `db.dialects` module
