+++ python/felis/db/utils.py
@@ -23,11 +23,10 @@ from __future__ import annotations
 
 import logging
 import re
-from types import ModuleType
 from typing import IO, Any
 
 import sqlalchemy
-from sqlalchemy import MetaData, dialects
+from sqlalchemy import MetaData
 from sqlalchemy.engine import Dialect, Engine, ResultProxy
 from sqlalchemy.engine.mock import MockConnection, create_mock_engine
 from sqlalchemy.engine.url import URL
@@ -36,50 +35,12 @@ from sqlalchemy.schema import CreateSchema, DropSchema
 from sqlalchemy.sql import text
 from sqlalchemy.types import TypeEngine
 
-from .sqltypes import MYSQL, ORACLE, POSTGRES, SQLITE
+from .dialects import get_dialect_module
 
-logger = logging.getLogger(__name__)
-
-_DIALECT_NAMES = [MYSQL, POSTGRES, SQLITE, ORACLE]
-
-
-def _dialect(dialect_name: str) -> Dialect:
-    """Create the SQLAlchemy dialect for the given name."""
-    return create_mock_engine(f"{dialect_name}://", executor=None).dialect
-
-
-def _dialect_module(dialect_name: str) -> ModuleType:
-    """Get the SQLAlchemy dialect module for the given name."""
-    return getattr(dialects, dialect_name)
-
-
-_DIALECTS = {name: _dialect(name) for name in _DIALECT_NAMES}
-"""Dictionary of dialect names to SQLAlchemy dialects."""
-
-_DIALECT_MODULES = {name: _dialect_module(name) for name in _DIALECT_NAMES}
-"""Dictionary of dialect names to SQLAlchemy modules for type instantiation."""
+logger = logging.getLogger("felis")
 
 _DATATYPE_REGEXP = re.compile(r"(\w+)(\((.*)\))?")
 """Regular expression to match data types in the form "type(length)"""
-
-
-def get_supported_dialects() -> dict[str, Dialect]:
-    """Get the support SQLAlchemy dialects."""
-    return _DIALECTS
-
-
-def get_dialect(dialect_name: str) -> Dialect:
-    """Get the SQLAlchemy dialect for the given name."""
-    if dialect_name not in get_supported_dialects():
-        raise ValueError(f"Unsupported dialect: {dialect_name}")
-    return get_dialect(dialect_name)
-
-
-def get_dialect_module(dialect_name: str) -> ModuleType:
-    """Get the SQLAlchemy dialect module for the given name."""
-    if dialect_name not in _DIALECT_MODULES:
-        raise ValueError(f"Unsupported dialect: {dialect_name}")
-    return _DIALECT_MODULES[dialect_name]
 
 
 def string_to_typeengine(
